Made small improvements at different places.
Added two formulas into tests::add_seq, which are used
in the main example of the calc.pdf document.
The formula can be transformed into NNF, but not further
because I need to think about the clause format first.

The formula initially had some type checking errors, and
I used this to improve error reporting in logic/structural.cpp