Added indexing operator at( ) to identifier.
In pretty, I created a separate function for printing
chains of identifiers with declarations.
It is used both by lambda and forall/exists.
It is possible that delta will also use it in the future.

Made changes in uniquenamestack. The older version always
removed indices, and put back its own indices. This makes
it impossible for the user to write forall( x1:O, x2:O) because
it becomes forall( x:O, x1:O ). The new version keeps the user
indices as long as they are strictly increasing.